Added Obstacles 5x4 grid
Added alternative Obstacles sheet with a 5x4 grid for holding 35x48mm Post-it notes.
You can prioritize or group together related obstacles in different rows and columns. 
There is a placeholder box to hold the one obstacle that you are working on now. It's only big enough to hold one sticker!